+++ funcionario.h
@@ -14,7 +14,7 @@ typedef struct funcionario Funcionario;
 /*Função funcionario
 Cadastra um novo arquivo funcionario, alocando memoria*/
 
-Funcionario * criarFuncionario (char * nome, char * cargo, int documento);
+Funcionario * criarFuncionario (char * nome, char * cargo,   long long int documento);
 
 /* FUnçao para liberar memoria alocada*/
 
@@ -33,7 +33,7 @@ void lerarquivo(FILE *fp, Funcionario * funcionarios[], int * count_fun);
 
 /* Funçao para buscar funcionario por documento*/
 
-Funcionario *buscaLineardocumento(int count_fun, Funcionario * funcionario[], int documento);
+Funcionario *buscaLineardocumento(int count_fun, Funcionario * funcionario[],   long long int documento);
 
 /*Função para buscar funcionario por nome*/
